requirements.txt
uvicorn[standard]
fastapi
pydantic
httpx
python-multipart
python-dotenv
PyMuPDF
pypdf
pdfplumber
